Set up a project for spine-c.
