+++ shared-src/spine_manipulator_dialogue.cpp
